+++ Changelog.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## Version 0.2.0
+- Created ProjectServlet.java. It is the start of my project
+- Included covid-19.csv file
+- Updated pom.xml
+
 ## Version 0.1.0
 - Tomcat and Apache Spark work